Split Double into Stub and Mock classes
Double is gone. You must now choose if to use a Stub or a Mock.

By default, Stub is not callable unless you explicitly stub a call to
it. This will avoid false positives in a situation like this:

    def create_record(validator, data):
        if validator.is_valid(data):
            # save the data...

lets say I'm testing it with a stub like this:

    calling(validator).passing('valid data').returns(True)
    calling(validator).passing('invalid data').returns(False)

    data = 'valid data'
    create_record(validator, data)

The test will pass. All good. But if stubs are callable by default, I
could change the implementation to look like this:

    # oops, forgot to pass the argument:
    if validator.is_valid():
        # save the data...

and my test will still pass! Making stubs not callable by default
removes that risk. As of this commit, the above scenario would raise a
TypeError: <Stub ...> is not callable.

Now let's say the body of that `if` block looks like this:

    repository.insert(data)

No logic depends on the return value, so stubbing the call in our test
before the execution step will distract from the story the tests are
trying to tell. So the repository should be replaced with a Mock, which
will be callable by default, and record calls for verification
(functionality that does not exist in Stubs).

This split also has the benefit of making you think harder about your
design. It will favor command-query-separation. Stubs will be useful for
queries. Mocks will be useful for commands.

Also closes #3 since I needed better reprs to distinguish between Stubs
and Mocks.