Configure nosetests to ignore tube.py
Because we're using the doctest plugin, nosetests will look in tube.py for docs with tests in them. That's okay in tdubs dev environments because testtube will be installed. However, when running python setup.py nosetests in a non-dev environment context, the tests command will fail because we don't list testtube in setup_requires.