setup.py: specify dependencies for all commands
In order for all of setup.py's commands to run correctly, we need setuptools to first install several dependencies if they aren't already present. Without these things installed, 1. the test command won't execute any tests. 2. the nosetests command won't be found if you try to use setup.py to run it and it will throw error if you try to execute it.

After these changes, one should be able to download the tdubs package and verify it in an empty virtual environment.