Correctly configure nose's ignore-files option
When one specifies an ignore-files configuration for nose, it clobbers the default config. This is apparently by design and is discussed here: https://github.com/nose-devs/nose/issues/872

One of the things the default config does is ignore setup.py. Not ignoring it causes really weird behavior including breaking the `nosetests` command and messing up coverage reports. This change amends the default ignore-files config so that it also ignores tube.py.